Workaround: NSNumber can not store unsigned values
NSNumber does not store unsigned values *as unsigned*.  When initialized with an unsigned number, NSNumber stores it as signed using the next highest bit-width.

This is a problem for Mach-O Kit’s description system, parts of which expect that the objCType of an NSNumber will exactly correspond to the primitive type that NSNumber was initialized with.

The workaround is to define a custom NSNumber subclass for storing unsigned values.  A new macro `_$()` is added to replace `@()` for boxing primitive numbers.  `_$()` uses the  MK NSNumber subclass where appropriate.