Add an alternateField key to MKNodeField
Alternate fields are used to augment the description of a field.  The alternate field names another field that is attached to the same node as the (host) field.  The alternate field description and value are used to augment the description of the host field in various ways (determined by the field flags).